suzerain: remedy signed/unsigned warning
git-svn-id: https://svn.ices.utexas.edu/repos/pecos/suzerain/trunk@41702 0bd85f0b-c151-43b9-a1e9-27231104551f

+++ apps/perfect/navier_stokes.hpp
@@ -752,7 +752,7 @@ std::vector<real_t> apply_navier_stokes_spatial_operator(
         rqq_y.col(ndx::rho) = common.rho  ();
 
         ArrayXr tmp;
-        for (size_t i = 0; i < rqq_y.cols(); ++i) {
+        for (int i = 0; i < rqq_y.cols(); ++i) {
             tmp = rqq_y.col(i);
             o.masslu()->solve(tmp.cols(), tmp.data(),
                               tmp.innerStride(), tmp.outerStride());
